Resolved issue regarding unresponsive Onboarding views.
