+++ iOS/Labs8-MealHelper/Labs8-MealHelper/AppDelegate.swift
@@ -15,7 +15,14 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
 
 
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
-        // Override point for customization after application launch.
+        
+        window = UIWindow(frame: UIScreen.main.bounds)
+        window?.makeKeyAndVisible()
+        let loginController = LoginViewController()
+        let navController = UINavigationController(rootViewController: loginController)
+        navController.navigationBar.prefersLargeTitles = true
+        window?.rootViewController = navController
+        
         return true
     }
