+++ iOS/Labs8-MealHelper/Labs8-MealHelper/Login/OnboardingCollectionViewCell.swift
@@ -14,14 +14,18 @@ protocol OnboardingCollectionViewCellDelegate {
     func save()
 }
 
-class OnboardingCollectionViewCell: UICollectionViewCell {
+class OnboardingCollectionViewCell: UICollectionViewCell, UITextFieldDelegate {
     
     // MARK: - Properties
+    
+    let genders = ["Female", "Male"]
+    let inputFields = ["Username", "Email", "Password", "Birthday", "Height", "Weight"]
+    var textFields = [UITextField]()
     
     var delegate: OnboardingCollectionViewCellDelegate?
     var isLastCell: Bool = false {
         didSet {
-            setupViews()
+            setupHeader()
         }
     }
     
@@ -31,15 +35,10 @@ class OnboardingCollectionViewCell: UICollectionViewCell {
         stackView.axis = .vertical
         stackView.distribution = .fillProportionally
         stackView.alignment = .fill
+        stackView.spacing = 30.0
         return stackView
     }()
     
-    private let headerView: UIView = {
-        let view = UIView()
-        view.translatesAutoresizingMaskIntoConstraints = false
-        return view
-    }()
-
     private let headerStackView: UIStackView = {
         let stackView = UIStackView()
         stackView.translatesAutoresizingMaskIntoConstraints = false
@@ -47,6 +46,21 @@ class OnboardingCollectionViewCell: UICollectionViewCell {
         stackView.distribution = .equalSpacing
         stackView.alignment = .center
         return stackView
+    }()
+    
+    private let genderButtonStackView: UIStackView = {
+        let view = UIStackView()
+        view.translatesAutoresizingMaskIntoConstraints = false
+        view.axis = .horizontal
+        view.distribution = .fillEqually
+        view.spacing = 8
+        return view
+    }()
+    
+    private let headerView: UIView = {
+        let view = UIView()
+        view.translatesAutoresizingMaskIntoConstraints = false
+        return view
     }()
     
     private let backButton: UIButton = {
@@ -108,73 +122,23 @@ class OnboardingCollectionViewCell: UICollectionViewCell {
     }
     
     @objc func save() {
+        print("save")
         delegate?.save()
+    }
+    
+    @objc func tappedGenderButton() {
+        print("beeep")
     }
     
     // MARK: - Private
     
     private func setupViews() {
+        setupHeader()
         
-        contentView.addSubview(headerView)
-        headerView.addSubview(headerStackView)
-        headerStackView.addArrangedSubview(backButton)
-        headerStackView.addArrangedSubview(headerTitleLabel)
-        headerStackView.addArrangedSubview(nextButton)
-        headerView.addSubview(mainStackView)
         mainStackView.addArrangedSubview(questionLabel)
-        let personalInfoView = PersonalInfoView()
-        mainStackView.addArrangedSubview(personalInfoView)
+        addSubview(mainStackView)
         
-        headerView.anchor(top: contentView.safeAreaLayoutGuide.topAnchor, leading: contentView.leadingAnchor, bottom: nil, trailing: contentView.trailingAnchor, padding: UIEdgeInsets(top: 15.0, left: 15.0, bottom: 0, right: 15.0), size: CGSize(width: 0.0, height: 50.0))
-        headerStackView.anchor(top: headerView.topAnchor, leading: headerView.leadingAnchor, bottom: headerView.bottomAnchor, trailing: headerView.trailingAnchor)
-        mainStackView.anchor(top: headerView.bottomAnchor, leading: contentView.leadingAnchor, bottom: contentView.bottomAnchor, trailing: contentView.trailingAnchor, padding: UIEdgeInsets(top: 15.0, left: 15.0, bottom: 200.0, right: 15.0))
         
-        let nextButtonAction = isLastCell ? #selector(save) : #selector(nextCell)
-        let nextButtonTitle = isLastCell ? "Save" : "Next"
-        nextButton.addTarget(self, action: nextButtonAction, for: .touchUpInside)
-        nextButton.setTitle(nextButtonTitle, for: .normal)
-        backButton.addTarget(self, action: #selector(previousCell), for: .touchUpInside)
-        
-    }
-    
-}
-
-
-class PersonalInfoView: UIView {
-    
-    let genders = ["Female", "Male"]
-    let inputFields = ["Username", "Email", "Birthday", "Height", "Weight"]
-    
-    private let mainStackView: UIStackView = {
-        let view = UIStackView()
-        view.translatesAutoresizingMaskIntoConstraints = false
-        view.axis = .vertical
-        view.distribution = UIStackView.Distribution.equalCentering
-        view.alignment = .fill
-        return view
-    }()
-    
-    private let genderButtonStackView: UIStackView = {
-        let view = UIStackView()
-        view.translatesAutoresizingMaskIntoConstraints = false
-        view.axis = .horizontal
-        view.distribution = .equalSpacing
-        view.alignment = .center
-        //view.spacing = 8
-        return view
-    }()
-    
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        setupViews()
-    }
-    
-    required init?(coder aDecoder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-    
-    private func setupViews() {
-        self.addSubview(mainStackView)
         mainStackView.addArrangedSubview(genderButtonStackView)
         
         genders.forEach { (gender) in
@@ -187,7 +151,30 @@ class PersonalInfoView: UIView {
             mainStackView.addArrangedSubview(textField)
         }
         
-        mainStackView.anchor(top: self.topAnchor, leading: self.leadingAnchor, bottom: self.bottomAnchor, trailing: self.trailingAnchor)
+        mainStackView.anchor(top: headerView.bottomAnchor, leading: leadingAnchor, bottom: bottomAnchor, trailing: trailingAnchor, padding: UIEdgeInsets(top: 15.0, left: 15.0, bottom: 150.0, right: 15.0))
+        
+    }
+    
+    private func setupHeader() {
+        addSubview(headerView)
+        headerView.addSubview(headerStackView)
+        headerStackView.addArrangedSubview(backButton)
+        headerStackView.addArrangedSubview(headerTitleLabel)
+        headerStackView.addArrangedSubview(nextButton)
+        //headerView.addSubview(mainStackView)
+        
+        headerView.anchor(top: safeAreaLayoutGuide.topAnchor, leading: self.leadingAnchor, bottom: nil, trailing: trailingAnchor, padding: UIEdgeInsets(top: 15.0, left: 15.0, bottom: 0, right: 15.0), size: CGSize(width: 0.0, height: 50.0))
+        headerStackView.anchor(top: headerView.topAnchor, leading: headerView.leadingAnchor, bottom: headerView.bottomAnchor, trailing: headerView.trailingAnchor)
+        
+        let nextButtonAction = isLastCell ? #selector(save) : #selector(nextCell)
+        let nextButtonTitle = isLastCell ? "Save" : "Next"
+        nextButton.addTarget(self, action: nextButtonAction, for: .touchUpInside)
+        nextButton.setTitle(nextButtonTitle, for: .normal)
+        backButton.addTarget(self, action: #selector(previousCell), for: .touchUpInside)
+    }
+    
+    func textFieldDidBeginEditing(_ textField: UITextField) {
+        
     }
     
     private func createButton(for gender: String) -> UIButton {
@@ -199,8 +186,9 @@ class PersonalInfoView: UIView {
         button.layer.masksToBounds = true
         button.backgroundColor = .red
         button.tintColor = .white
-        button.widthAnchor.constraint(equalToConstant: 100.0).isActive = true
+        //button.widthAnchor.constraint(equalToConstant: 100.0).isActive = true
         button.heightAnchor.constraint(equalToConstant: 50.0).isActive = true
+        button.addTarget(self, action: #selector(tappedGenderButton), for: .touchUpInside)
         return button
     }
     
@@ -218,7 +206,97 @@ class PersonalInfoView: UIView {
         tf.contentHorizontalAlignment = .left
         tf.heightAnchor.constraint(equalToConstant: 40.0).isActive = true
         tf.isUserInteractionEnabled = true
+        tf.delegate = self
         return tf
+        
     }
     
 }
+
+//
+//class PersonalInfoView: UIView {
+//
+//    let genders = ["Female", "Male"]
+//    let inputFields = ["Username", "Email", "Password", "Birthday", "Height", "Weight"]
+//
+//    let mainStackView: UIStackView = {
+//        let view = UIStackView()
+//        view.translatesAutoresizingMaskIntoConstraints = false
+//        view.axis = .vertical
+//        view.distribution = UIStackView.Distribution.equalCentering
+//        view.alignment = .fill
+//        return view
+//    }()
+//
+//    private let genderButtonStackView: UIStackView = {
+//        let view = UIStackView()
+//        view.translatesAutoresizingMaskIntoConstraints = false
+//        view.axis = .horizontal
+//        view.distribution = .fillEqually
+//        view.spacing = 8
+//        return view
+//    }()
+//
+//    override init(frame: CGRect) {
+//        super.init(frame: frame)
+//        setupViews()
+//    }
+//
+//    required init?(coder aDecoder: NSCoder) {
+//        fatalError("init(coder:) has not been implemented")
+//    }
+//
+//    private func setupViews() {
+//        self.addSubview(mainStackView)
+//        mainStackView.addArrangedSubview(genderButtonStackView)
+//
+//        genders.forEach { (gender) in
+//            let button = createButton(for: gender)
+//            genderButtonStackView.addArrangedSubview(button)
+//        }
+//
+//        inputFields.forEach { (input) in
+//            let textField = createTextField(for: input)
+//            mainStackView.addArrangedSubview(textField)
+//        }
+//
+//        mainStackView.anchor(top: self.topAnchor, leading: self.leadingAnchor, bottom: self.bottomAnchor, trailing: self.trailingAnchor)
+//    }
+//
+//    @objc func tappedGenderButton() {
+//        print("beeep")
+//    }
+//
+//    private func createButton(for gender: String) -> UIButton {
+//        let button = UIButton(type: .system)
+//        button.translatesAutoresizingMaskIntoConstraints = false
+//        button.setTitle(gender, for: .normal)
+//        button.titleLabel?.font = UIFont.systemFont(ofSize: 20.0)
+//        button.layer.cornerRadius = 8.0
+//        button.layer.masksToBounds = true
+//        button.backgroundColor = .red
+//        button.tintColor = .white
+//        //button.widthAnchor.constraint(equalToConstant: 100.0).isActive = true
+//        button.heightAnchor.constraint(equalToConstant: 50.0).isActive = true
+//        button.addTarget(self, action: #selector(tappedGenderButton), for: .touchUpInside)
+//        return button
+//    }
+//
+//    private func createTextField(for input: String) -> UITextField {
+//        let tf = UITextField()
+//        tf.translatesAutoresizingMaskIntoConstraints = false
+//        tf.placeholder = input
+//        tf.font = UIFont.systemFont(ofSize: 17.0)
+//        tf.setLeftPaddingPoints(10.0)
+//        tf.setRightPaddingPoints(10.0)
+//        tf.backgroundColor = .white
+//        tf.tintColor = .blue
+//        tf.layer.cornerRadius = 8.0
+//        tf.layer.masksToBounds = true
+//        tf.contentHorizontalAlignment = .left
+//        tf.heightAnchor.constraint(equalToConstant: 40.0).isActive = true
+//        tf.isUserInteractionEnabled = true
+//        return tf
+//    }
+//
+//}
